# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.StockMovement:
    properties:
      actor_id:
        type: integer
      actor_name:
        type: string
      category:
        type: string
      created_at:
        type: string
      field_crop_id:
        type: integer
      field_operation_id:
        type: integer
      field_operation_label:
        type: string
      id:
        type: integer
      movement_type:
        $ref: '#/definitions/domain.StockMovementType'
      notes:
        type: string
      quantity_delta:
        type: number
      resource_id:
        type: integer
      resource_name:
        type: string
      resulting_quantity:
        type: number
      total_cost:
        type: number
      unit:
        type: string
      unit_cost:
        type: number
    type: object
  domain.StockMovementType:
    enum:
    - in
    - out
    - adjustment
    type: string
    x-enum-varnames:
    - StockMovementIn
    - StockMovementOut
    - StockMovementAdjustment
...
  handlers.createStockMovementRequest:
    properties:
      field_operation_id:
        type: integer
      movement_type:
        enum:
        - in
        - out
        - adjustment
        type: string
      notes:
        type: string
      quantity:
        type: number
      resource_id:
        type: integer
      unit_cost:
        type: number
    required:
    - movement_type
    - resource_id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API REST pentru managementul mașinilor agricole, operatorilor și asignărilor.
  title: Agri API
  version: "1.0"
paths:
...
  /stock-movements:
    get:
      parameters:
      - description: Filtru resursă
        in: query
        name: resource_id
        type: integer
      - description: Filtru operațiune pe teren
        in: query
        name: field_operation_id
        type: integer
      - description: De la (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Până la (YYYY-MM-DD, inclusiv)
        in: query
        name: to
        type: string
      - description: Număr maxim (implicit 200, maxim 500)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.StockMovement'
            type: array
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mișcări de stoc
      tags:
      - stocks
    post:
      consumes:
      - application/json
      parameters:
      - description: Mișcarea de stoc
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.createStockMovementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.StockMovement'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Înregistrare mișcare de stoc
      tags:
      - stocks
...
securityDefinitions:
  BearerAuth:
    description: 'Introdu token-ul cu prefixul "Bearer ": **Bearer &lt;token&gt;**'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"